pyproject.toml
[project]
name = "portfoliooptimizator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.5.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "yfinance>=1.0",
]
